add/improve English translation
I made a pretty straightforward translation for the English version, with some small changes to break up the rhythm in the opening paragraph. Two sentences which I am not so sure about are: 

> Son un espacio a deconstruir.

and 

> Gestionando la habilitación para que los cursos y talleres otorguen puntaje docente.

You may want to take a closer look at the translation and see if it needs any changes!